+++ src/OpenLoco/src/Config.h
@@ -161,6 +161,7 @@ namespace OpenLoco::Config
         bool trainsReverseAtSignals = true;
         bool disableStationSizeLimit = false;
         bool showAiPlanningAsGhosts = false;
+        bool keepCargoModifyPickup = false;
 
         bool usePreferredOwnerName = false;
         std::string preferredOwnerName;